+++ src/components/templates/Page.template.tsx
@@ -103,7 +103,6 @@ export default async function [PageName]Page({ params, searchParams }: PageProps
 //   return (
 //     <main>
 //       {/* TODO: <VideoDetailHero video={video} /> */}
-//       {/* TODO: <VideoDetailMeta video={video} /> */}
 //       {/* TODO: <EpisodeGrid video={video} /> */}
 //     </main>
 //   )